Add changes to implement security and enable eureka and other minimal changes

+++ src/main/java/com/kyc/bills/entity/BillEntity.java
@@ -27,7 +27,7 @@ public class BillEntity {
     private Double total;
 
     @Column(value = "ID_CUSTOMER")
-    private Integer idCustomer;
+    private Long idCustomer;
 
     @Column(value = "SETTLED")
     private boolean settled;
